+++ CHANGELOG.md
@@ -34,6 +34,7 @@
 - Remove PackageKit in favor of bauh.
 - Add full support for ASUS laptops.
 - Add support for the Steam Deck OLED.
+- Change the display manager from LightDM to SDDM.
 
 ## 4.0.0 - The Conversion Update